// G7/pizzalina/server/src/shared/types/pizza-ingredient.enum.ts
export enum PizzaIngredientType {
  CHEESE = 'cheese',
  TOMATO = 'tomato',
  PEPPERONI = 'pepperoni',
  MUSHROOM = 'mushroom',
  OLIVES = 'olives',
  ONION = 'onion',
  BACON = 'bacon',
  HAM = 'ham',
  PINEAPPLE = 'pineapple',
  PEPPERS = 'peppers',
  CHICKEN = 'chicken',
  SPINACH = 'spinach',
  GARLIC = 'garlic',
  SAUSAGE = 'sausage',
  BEEF = 'beef',
  ANCHOVY = 'anchovy',
  TUNA = 'tuna',
  CORN = 'corn',
  BROCCOLI = 'broccoli',
  ARTICHOKE = 'artichoke',
  EGG = 'egg',
  PARMESAN = 'parmesan',
  MOZZARELLA = 'mozzarella',
  RICOTTA = 'ricotta',
  BLUE_CHEESE = 'blue_cheese',
  ROCKET = 'rocket',
  ZUCCHINI = 'zucchini',
  EGGPLANT = 'eggplant',
  PROSCIUTTO = 'prosciutto',
  SALAMI = 'salami',
  BBQ_SAUCE = 'bbq_sauce',
  BUFFALO_SAUCE = 'buffalo_sauce',
  BASIL = 'basil',
  OREGANO = 'oregano',
  // Add more as needed
}